cantidad en pedido representada por tinyint

+++ database/migrations/2022_04_24_161622_create_pedido_productos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('pedido_id');             
             $table->unsignedBigInteger('producto_id');            
-            $table->integer('cantidad');
+            $table->tinyInteger('cantidad');
         });
     }
